+++ power/__init__.py
@@ -1,0 +1,5 @@
+"""
+Utilities to handle the output of Pencil's power_xy subroutine.
+"""
+
+from . import cached