+++ read.py
@@ -28,7 +28,7 @@ import abc
 
 from dataclasses import dataclass
 
-from .power_cached import read_power
+from .power.cached import read_power
 from .utils import smooth_tophat
 
 class plot_container():